requirements-dev.txt
pycodestyle==2.8.0
pytest==7.1.1
zipp==3.7.0
pytest-django==4.5.2
autopep8==1.6.0
pylint==2.12.1
mypy==0.942
black==22.3.0
django-stubs==1.10.1
types-ujson==4.2.1